+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `opentelemetry-instrumentation-aws-lambda`:  Implement new semantic convention opt-in with stable http semantic conventions
+  ([#3939](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3939))
+
 ## Version 1.40.0/0.61b0 (2026-03-04)
 
 ### Added